// src/app/models/estabelecimento.ts
export class Estabelecimento
{
  key: string;
  IdEstabelecimento: string;
  NomeEstabelecimento: string;
  BloqueioFila: string;
  PausaFilsa: string;
  Celular: string;
  Email: string;
  PossuiFila: string;
  TempoMedio: BigInteger;
  TamanhoMaximoFila: BigInteger;
  Senha: String;
  keyfila: String;
}